move agent modules into weave/trace_server/agents/
Per griffin's review: group the agent-related source files under a
common subfolder instead of sprinkling agent_*.py across trace_server/.

- agent_chat_view.py    -> agents/chat_view.py
- agent_clickhouse.py   -> agents/clickhouse.py
- agent_helpers.py      -> agents/helpers.py
- agent_schema.py       -> agents/schema.py
- agent_types.py        -> agents/types.py
- semconv.py            -> agents/semconv.py

The SQL query builder stays under query_builder/ next to its peers
(annotation_queues_query_builder, objects_query_builder, etc.) so it's
easy to find alongside other query builders.
