+++ weave/trace_server/trace_server_interface.py
@@ -15,7 +15,7 @@ else:
         ResourceSpans = Any
 
 from weave.trace_server import http_service_interface as his
-from weave.trace_server.agent_types import *  # noqa: F403
+from weave.trace_server.agents.types import *  # noqa: F403
 from weave.trace_server.common_interface import (
     WB_USER_ID_DESCRIPTION,
     AnnotationState,
